Add: adding file

+++ 0x1E-search_algorithms/0-linear.c
@@ -15,14 +15,14 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i = 0;
+	unsigned int i = 0;
 
 	if (array == NULL)
 		return (-1);
 
 	while (i < size)
 	{
-		printf("Value checked array [%lu] = [%d]\n", i, array[i]);
+		printf("Value checked array [%u] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 		i++;
